add superclass BigIntReader and extend to UInt64Reader and UInt128Reader
