+++ settings.gradle
@@ -2,11 +2,12 @@ rootProject.name = 'polkaj-root'
 
 include "polkaj-ss58",
         "polkaj-scale",
-        // "polkaj-json-types",
+        "polkaj-json-types",
         "polkaj-scale-types",
-        // "polkaj-schnorrkel",
-        "polkaj-common-types"
-        // "polkaj-api-base",
-        // "polkaj-api-http",
-        // "polkaj-api-ws",
-        // "polkaj-tx"
+        "polkaj-schnorrkel",
+        "polkaj-common-types",
+        "polkaj-api-base",
+        "polkaj-api-http",
+        "polkaj-api-ws",
+        "polkaj-tx"
+
